Make more setup/behavior optional.
This adds three new class options

    [no]bib
    [no]frontmatter
    [no]hyperref

all of which disable the corresponding, relevant features. The
motivation is that something like TikZ + the preview package should
be able to setup the environment with the same/similar preamble as
the main document, but skip a bunch of unnecessary setup which won't
be needed in such a case.